Inheritance
Weapons now inherit from a base class "Weapon".

This class has variables for current number and max amount of attacks, the function that says to attack if left mouse button has been clicked, a counter for current attacks, as well as the function that destroys the gameObject if it has exceeded this max amount.