Recovery stuff (See description)
Some of the stuff that was committed to a branch that dissapeared.
Created a copy of the game scene for testing. Copied the enemy script to test and also use as the new and improved script in the future. Created a new sprite/model that the enemy will use. Incorperated the inventory so the enemy will use the weapon in its hand. The enemy aiming is not fixed, as Elias is working on that, and I still need to change some of the comments, so it matches the missing branch